src/requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.15.2
drf-nested-routers==0.94.1
drf-yasg==1.21.9
inflection==0.5.1
packaging==24.2
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
